Fixing some test handling of symbol font
