+++ internal/painter/font_testhelper.go
@@ -1,0 +1,9 @@
+//go:build ci || test
+
+package painter
+
+import "github.com/go-text/typesetting/fontscan"
+
+func loadSystemFonts(fm *fontscan.FontMap) error {
+	return nil
+}
